refactor: reorder component and move to components folder
